Added a helper library for my functions, added in picture and date for recipes
